refactor: rename ConsentManager to ConsentService and move consent.ts to domain for consistency.
